Packet_analyzer/requirements.txt
requests
numpy
scipy
python-dotenv
pandas
joblib
psutil
scikit-learn
xgboost
